support for round robin allocation of cuda cards to workers
A gunicorn post_fork hook has been added to set CUDA_VISIBLE_DEVICES, which
sets the device torch will use.

A app level config variable "APP_CUDA_DEVICE_COUNT" is required to
indicate how many devices are to be used.

The devices are allocated to the docker in the docker compose configuration.